// rust-port/wifi-densepose-rs/crates/wifi-densepose-desktop/ui/src/types.ts
export interface AppSettings {
  server_http_port: number;
  server_ws_port: number;
  server_udp_port: number;
  bind_address: string;
  ui_path: string;
  ota_psk: string;
  auto_discover: boolean;
  discover_interval_ms: number;
  theme: "dark" | "light";
}